Implemented HealthCheck Endpoint and improved Mailman resilience
The /v1/moc/HealthCheck endpoint will perform checks for various
external components that the service needs access to. Ex. SSH
connection to the mailing list server, which is the primary
source of flakiness at the moment.

Additionally, refactored the Mailman code into a separate class, to
facilitate testing and health checking.

Finally, made failing to subscribe a user to a mailing list not
be a blocking error.